Reduce hourly backups to 6 days; allow override
Make sure we can override the default number of days that we
keep backups for both daily and hourly.

It is however impractical to store huge number of days of hourly
backups (due to the number of files to sync with at the remote). An
hourly backups usefulness deminishes over time, so only store 6 days
by default.

*** It is vital that the backup area covered by the hourly backup is
    also covered by the daily backup for this logic to make sense ***