Handle pgsql backup from remote host.
Adds the following configuration parameters to pgsql backupninja
helper:

* user
* password
* host
* port (default 5432)

In short, adds remote database support to the pgsql helper.

This support is added via the use of PGHOST, PGUSER, PGPORT, and via
creation of a tempfile in .pgpass format, spec'd via PGPASSFILE.

This means that we cannot run the backups via the same 'su to PGSQLUSER'
process or under the vserver processes as before, though in fairness when
dealing with remote databases these are redundant anyway. It definitely keeps
the code cleaner.

I've added '-w' to all psql and pg_dump* commands, so that we never get asked for
a password. This should prevent a potential hang.